+++ src/db/Migration.php
@@ -195,12 +195,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function insert($table, $columns, bool $includeAuditColumns = true): void
     {
-        echo "    > insert into $table ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("insert into $table");
         $this->db->createCommand()
             ->insert($table, $columns, $includeAuditColumns)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     /**
@@ -215,12 +214,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function batchInsert($table, $columns, $rows, bool $includeAuditColumns = true): void
     {
-        echo "    > batch insert into $table ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("batch insert into $table");
         $this->db->createCommand()
             ->batchInsert($table, $columns, $rows, $includeAuditColumns)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     /**
@@ -256,12 +254,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function update($table, $columns, $condition = '', $params = [], bool $includeAuditColumns = true): void
     {
-        echo "    > update in $table ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("update in $table");
         $this->db->createCommand()
             ->update($table, $columns, $condition, $params, $includeAuditColumns)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     /**
@@ -292,12 +289,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function replace(string $table, string $column, string $find, string $replace, $condition = '', array $params = []): void
     {
-        echo "    > replace \"$find\" with \"$replace\" in $table.$column ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("replace \"$find\" with \"$replace\" in $table.$column");
         $this->db->createCommand()
             ->replace($table, $column, $find, $replace, $condition, $params)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     // Schema Manipulation Methods
@@ -310,12 +306,66 @@ abstract class Migration extends \yii\db\Migration
      */
     public function dropTableIfExists(string $table): void
     {
-        echo "    > dropping $table if it exists ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("dropping $table if it exists");
         $this->db->createCommand()
             ->dropTableIfExists($table)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
+    }
+
+    /**
+     * Creates and executes a SQL statement for dropping an index if it exists.
+     *
+     * @param string $table The table that the index was created for. The table name will be properly quoted by the method.
+     * @param string|string[] $columns The column(s) that are included in the index. If there are multiple
+     * columns, separate them by commas or use an array.
+     * @param bool $unique Whether the index has a UNIQUE constraint.
+     * @since 4.0.0
+     */
+    public function dropIndexIfExists(string $table, $columns, bool $unique = false): void
+    {
+        $time = $this->beginCommand("dropping index on $table if it exists");
+        Db::dropIndexIfExists($table, $columns, $unique, $this->db);
+        $this->endCommand($time);
+    }
+
+    /**
+     * Creates and executes a SQL statement for dropping a foreign key if it exists.
+     *
+     * @param string $table The table that the foreign key was created for. The table name will be properly quoted by the method.
+     * @param string|string[] $columns The column(s) that are included in the foreign key. If there are multiple
+     * columns, separate them by commas or use an array.
+     * @since 4.0.0
+     */
+    public function dropForeignKeyIfExists(string $table, $columns): void
+    {
+        $time = $this->beginCommand("dropping foreign key on $table if it exists");
+        Db::dropForeignKeyIfExists($table, $columns, $this->db);
+        $this->endCommand($time);
+    }
+
+    /**
+     * Creates and executes a SQL statement for dropping all foreign keys to a table.
+     *
+     * @param string $table The table that the foreign keys should reference.
+     * @since 4.0.0
+     */
+    public function dropAllForeignKeysToTable(string $table): void
+    {
+        $time = $this->beginCommand("dropping all foreign keys to $table");
+        Db::dropAllForeignKeysToTable($table, $this->db);
+        $this->endCommand($time);
+    }
+
+    /**
+     * Builds and executes a SQL statement for renaming a DB table and its corresponding sequence (if PostgreSQL).
+     * @since 4.0.0
+     */
+    public function renameTable($table, $newName)
+    {
+        $time = $this->beginCommand("rename table $table to $newName");
+        Db::renameTable($table, $newName);
+        $this->endCommand($time);
     }
 
     /**
@@ -326,12 +376,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function renameSequence(string $oldName, string $newName): void
     {
-        echo "    > rename sequence $oldName to $newName ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("rename sequence $oldName to $newName");
         $this->db->createCommand()
             ->renameSequence($oldName, $newName)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     /**
@@ -397,12 +446,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function softDelete(string $table, $condition = '', array $params = []): void
     {
-        echo "    > soft delete from $table ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("soft delete from $table");
         $this->db->createCommand()
             ->softDelete($table, $condition, $params)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     /**
@@ -416,12 +464,11 @@ abstract class Migration extends \yii\db\Migration
      */
     public function restore(string $table, $condition = '', array $params = []): void
     {
-        echo "    > restore from $table ...";
-        $time = microtime(true);
+        $time = $this->beginCommand("restore from $table");
         $this->db->createCommand()
             ->restore($table, $condition, $params)
             ->execute();
-        echo ' done (time: ' . sprintf('%.3f', microtime(true) - $time) . "s)\n";
+        $this->endCommand($time);
     }
 
     /**